[Feat] [GGFE-145] 아이템 리스트 조회 mock api 적용
